+++ CHANGELOG.md
@@ -2,7 +2,17 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-## [2.3.0](https://github.com/ObservaSolutions/sdk-observa/compare/v2.2.4...v2.3.0) (2026-02-16)
+## [2.4.0] (2026-02-19)
+
+### Features
+* Support for browser environments using `publicKey`.
+* Allow SDK initialization without `apiKey` if `publicKey` is provided.
+* Add `publicKey` to ingest events and health check payloads.
+
+### Notes
+* Mobile support is planned for future releases.
+
+### [2.3.0](https://github.com/ObservaSolutions/sdk-observa/compare/v2.2.4...v2.3.0) (2026-02-16)
 
 ### [2.2.4](https://github.com/ObservaSolutions/sdk-observa/compare/v2.2.3...v2.2.4) (2026-02-15)